refactor!: Rename src to core and move core service out of the services
BREAKING CHANGE: src folder has been renamed to core, update your imports accordingly
