+++ projects/hslayers-app/src/hslayers-app/hslayers-app.module.ts
@@ -2,7 +2,7 @@ import {ApplicationRef, DoBootstrap, NgModule} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
 
 import {HslayersAppComponent} from './hslayers-app.component';
-import {HslayersModule} from 'hslayers-ng';
+import {HslayersModule} from 'hslayers-ng/core';
 
 @NgModule({
   declarations: [HslayersAppComponent],
