+++ projects/hslayers-cesium-app/src/app/app.module.ts
@@ -3,7 +3,7 @@ import {BrowserModule} from '@angular/platform-browser';
 
 import {AppComponent} from './app.component';
 import {HsCesiumModule} from 'hslayers-cesium';
-import {HslayersModule} from 'hslayers-ng';
+import {HslayersModule} from 'hslayers-ng/core';
 
 @NgModule({
   declarations: [AppComponent],
