--- projects/hslayers/components/map/index.ts
@@ -1,1 +1,0 @@
-export * from './public-api';
