--- projects/hslayers/components/map/public-api.ts
@@ -1,3 +1,0 @@
-export * from './map.component';
-export * from './map.directive';
-export * from './map.module';
