+++ projects/hslayers/core/hslayers.component.scss
