+++ projects/hslayers/core/hslayers.component.ts
