+++ projects/hslayers/core/layout.directive.ts
