+++ projects/hslayers/core/map/map.component.html
