+++ projects/hslayers/core/map/map.directive.ts
@@ -2,6 +2,7 @@ import {Directive, ViewContainerRef} from '@angular/core';
 
 @Directive({
   selector: '[map]',
+  standalone: true,
 })
 export class HsMapDirective {
   constructor(public viewContainerRef: ViewContainerRef) {}
