+++ projects/hslayers/core/ng-package.json
