+++ projects/hslayers/core/public-api.ts
@@ -5,3 +5,5 @@ export * from './hslayers.component';
 export * from './hslayers.module';
 export * from './map-host.directive';
 export * from './layout.directive';
+export * from './core.service';
+export * from './map/map.component';
