+++ projects/hslayers/css/hslayers-bootstrap.scss
@@ -980,7 +980,7 @@ $list-group-item-padding-x: 1.25rem;
     }
   }
 
-  @import "../src/hslayers.component.scss";
+  @import "../core/hslayers.component.scss";
   @import "../components/layer-manager/layer-manager.component.scss";
   @import "../components/layer-manager/logical-list/layerlist.component.scss";
   @import "../components/layer-manager/gallery/layer-manager-gallery.component.scss";
