+++ projects/hslayers/public-api.ts
@@ -1,2 +1,4 @@
-//Dummy variable export to 'fool' compiler
-export const HSLAYERSNG = 'hslayers-ng';
+// public-api.ts is required by ng-packagr and cannot be empty.
+const DO_NOT_IMPORT =
+  'Do not import from hslayers-ng; Submodules must be imported directly.';
+export default DO_NOT_IMPORT;
