+++ projects/hslayers/services/sidebar/sidebar.service.ts
@@ -3,7 +3,7 @@ import {Injectable} from '@angular/core';
 
 import {HsButton} from 'hslayers-ng/types';
 import {HsConfig} from 'hslayers-ng/config';
-import {HsCoreService} from 'hslayers-ng/services/core';
+import {HsCoreService} from 'hslayers-ng/core';
 import {HsLanguageService} from 'hslayers-ng/services/language';
 import {HsLayoutService} from 'hslayers-ng/services/layout';
 import {HsUtilsService} from 'hslayers-ng/services/utils';
