+++ projects/test-app/src/hslayers-app/hslayers-app.module.ts
@@ -2,7 +2,7 @@ import {BrowserModule} from '@angular/platform-browser';
 import {NgModule} from '@angular/core';
 
 import {HslayersAppComponent} from './hslayers-app.component';
-import {HslayersModule} from 'hslayers-ng';
+import {HslayersModule} from 'hslayers-ng/core';
 import {SomeModule} from './some-panel/some-panel.module';
 
 @NgModule({
